+++ crates/holochain_state/src/query/link_details.rs
@@ -23,9 +23,9 @@ impl GetLinkDetailsQuery {
 }
 
 impl Query for GetLinkDetailsQuery {
-    type Item = Judged<SignedHeaderHashed>;
-    type State = HashMap<HeaderHash, (Option<SignedHeaderHashed>, HashSet<SignedHeaderHashed>)>;
-    type Output = Vec<(SignedHeaderHashed, Vec<SignedHeaderHashed>)>;
+    type Item = Judged<SignedActionHashed>;
+    type State = HashMap<ActionHash, (Option<SignedActionHashed>, HashSet<SignedActionHashed>)>;
+    type Output = Vec<(SignedActionHashed, Vec<SignedActionHashed>)>;
     fn query(&self) -> String {
         self.query.query()
     }
@@ -39,7 +39,7 @@ impl Query for GetLinkDetailsQuery {
     }
 
     fn as_map(&self) -> Arc<dyn Fn(&Row) -> StateQueryResult<Self::Item>> {
-        let f = row_blob_to_header("header_blob");
+        let f = row_blob_to_action("action_blob");
         // Data is valid because it is filtered in the sql query.
         Arc::new(move |row| Ok(Judged::valid(f(row)?)))
     }
@@ -49,8 +49,8 @@ impl Query for GetLinkDetailsQuery {
         let base_filter = query.base.clone();
         let type_query_filter = query.type_query.clone();
         let tag_filter = query.tag.clone();
-        let f = move |header: &QueryData<Self>| match header.header() {
-            Header::CreateLink(CreateLink {
+        let f = move |action: &QueryData<Self>| match action.action() {
+            Action::CreateLink(CreateLink {
                 base_address,
                 tag,
                 link_type,
@@ -64,7 +64,7 @@ impl Query for GetLinkDetailsQuery {
                         .as_ref()
                         .map_or(true, |t| LinksQuery::tag_to_hex(tag).starts_with(&(**t)))
             }
-            Header::DeleteLink(DeleteLink { base_address, .. }) => *base_address == *base_filter,
+            Action::DeleteLink(DeleteLink { base_address, .. }) => *base_address == *base_filter,
             _ => false,
         };
         Box::new(f)
@@ -72,22 +72,22 @@ impl Query for GetLinkDetailsQuery {
 
     fn fold(&self, mut state: Self::State, data: Self::Item) -> StateQueryResult<Self::State> {
         let shh = data.data;
-        let header = shh.header();
-        match header {
-            Header::CreateLink(_) => {
+        let action = shh.action();
+        match action {
+            Action::CreateLink(_) => {
                 state
                     .entry(shh.as_hash().clone())
                     .or_insert((Some(shh), HashSet::new()));
             }
-            Header::DeleteLink(delete_link) => {
+            Action::DeleteLink(delete_link) => {
                 let entry = state
                     .entry(delete_link.link_add_address.clone())
                     .or_insert((None, HashSet::new()));
                 entry.1.insert(shh);
             }
             _ => {
-                return Err(StateQueryError::UnexpectedHeader(
-                    shh.header().header_type(),
+                return Err(StateQueryError::UnexpectedAction(
+                    shh.action().action_type(),
                 ))
             }
         }
@@ -106,12 +106,12 @@ impl Query for GetLinkDetailsQuery {
             .filter_map(|(_, (create, deletes))| {
                 create.map(|create| {
                     let mut deletes = deletes.into_iter().collect::<Vec<_>>();
-                    deletes.sort_by_key(|l| l.header().timestamp());
+                    deletes.sort_by_key(|l| l.action().timestamp());
                     (create, deletes.into_iter().collect())
                 })
             })
             .collect::<Vec<_>>();
-        r.sort_by_key(|l| l.0.header().timestamp());
+        r.sort_by_key(|l| l.0.action().timestamp());
         Ok(r)
     }
 }